refactor: rename telegram_interface to telegram_bot and update instructions
